SUBMARINE-712. [WEB] Refactor environment page of Workbench
