+++ submarine-workbench/workbench-web/src/app/pages/workbench/environment/environment-home/environment-home.component.html
@@ -1,1 +1,38 @@
-<p>environment-home works!</p>
+<!--
+  ~ Licensed to the Apache Software Foundation (ASF) under one
+  ~ or more contributor license agreements.  See the NOTICE file
+  ~ distributed with this work for additional information
+  ~ regarding copyright ownership.  The ASF licenses this file
+  ~ to you under the Apache License, Version 2.0 (the
+  ~ "License"); you may not use this file except in compliance
+  ~ with the License.  You may obtain a copy of the License at
+  ~
+  ~   http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing,
+  ~ software distributed under the License is distributed on an
+  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+  ~ KIND, either express or implied.  See the License for the
+  ~ specific language governing permissions and limitations
+  ~ under the License.
+  -->
+
+<div style="margin: 15px; padding: 15px; background-color: white">
+  <div align="right">
+    <button
+      nz-button
+      id="btn-newEnvironment"
+      nzType="primary"
+      style="margin: 10px 4px 10px 4px"
+      (click)="form.initModal()"
+    >
+      <i nz-icon nzType="plus"></i>
+      New Environment
+    </button>
+  </div>
+  <submarine-environment-list
+    [environmentList]="environmentList"
+    (deleteEnvironment)="onDeleteEnvironment($event)"
+  ></submarine-environment-list>
+  <submarine-environment-form #form (updater)="updateEnvironmentList($event)"></submarine-environment-form>
+</div>
