--- submarine-workbench/workbench-web/src/app/pages/workbench/environment/environment-info/environment-info.component.html
@@ -1,1 +1,0 @@
-<p>environment-info works!</p>
