--- submarine-workbench/workbench-web/src/app/pages/workbench/environment/environment-info/environment-info.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'submarine-environment-info',
-  templateUrl: './environment-info.component.html',
-  styleUrls: ['./environment-info.component.scss'],
-})
-export class EnvironmentInfoComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit() {}
-}
