+++ submarine-workbench/workbench-web/src/app/pages/workbench/environment/environment.component.scss
@@ -16,83 +16,3 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-#environmentOuter{
-  background-color: white;
-  padding-left: 30px;
-  padding-top: 20px;
-}
-
-#environmentData{
-  margin-top: 16px;
-  margin-left: 25px;
-  margin-right: 25px;
-  background-color:white;
-  padding-left: 10px;
-  padding-right: 10px;
-}
-
-input.ng-invalid.ng-touched {
-  border: 1px solid red;
-}
-
-textarea.ng-invalid.ng-touched {
-  border: 1px solid red;
-}
-
-.red-star {
-  margin-top: 20px;
-  color: red;
-}
-
-.list-container {
-  padding: 0;
-  margin: 1rem 0;
-}
-
-.input-group {
-  display: flex;
-  align-items: center;
-  &:not(:first-child) {
-    margin-top: 1rem;
-  }
-
-  & input {
-    flex: 0 2 20%;
-    &:not(:last-child) {
-      margin-right: .5rem;
-    }
-  }
-
-  & i {
-    cursor: pointer;
-    font-size: 1.2rem;
-  }
-}
-
-/* utility */
-.single-field-group {
-  display: flex;
-  align-items: center;
-  margin-bottom: 1.5rem;
-  & label {
-    flex: 0 0 25%;
-    text-align: right;
-    font-weight: 500;
-    &:not(:last-child) {
-      margin-right: 1.5rem;
-    }
-  }
-  & input {
-    flex: 0 0 48%;
-  }
-  & textarea {
-    flex: 0 0 55%;
-  }
-}
-
-.alert-message {
-  color: red;
-  margin-top: .3rem;
-  margin-left: .5rem;
-}
