+++ submarine-workbench/workbench-web/src/app/pages/workbench/environment/environment.component.ts
@@ -18,11 +18,6 @@
  */
 
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormArray, FormControl, FormGroup, Validators } from '@angular/forms';
-import { Environment } from '@submarine/interfaces/environment-interfaces/environment-info';
-import { EnvironmentService } from '@submarine/services/environment-services/environment.service';
-import { NzMessageService } from 'ng-zorro-antd';
-import { ExperimentValidatorService } from '@submarine/services/experiment.validator.service';
 
 @Component({
   selector: 'submarine-environment',
@@ -30,168 +25,7 @@ import { ExperimentValidatorService } from '@submarine/services/experiment.valid
   styleUrls: ['./environment.component.scss'],
 })
 export class EnvironmentComponent implements OnInit {
-  constructor(
-    private environmentService: EnvironmentService,
-    private nzMessageService: NzMessageService,
-    private fb: FormBuilder,
-    private experimentValidatorService: ExperimentValidatorService
-  ) {}
+  constructor() {}
 
-  environmentList: Environment[] = [];
-  checkedList: boolean[] = [];
-  selectAllChecked: boolean = false;
-
-  isVisible = false;
-  environmentForm;
-
-  ngOnInit() {
-    this.environmentForm = this.fb.group({
-      environmentName: [null, Validators.required],
-      dockerImage: [null, Validators.required],
-      name: [null, Validators.required],
-      channels: this.fb.array([]),
-      dependencies: this.fb.array([]),
-    });
-    this.fetchEnvironmentList();
-  }
-
-  fetchEnvironmentList() {
-    this.environmentService.fetchEnvironmentList().subscribe((list) => {
-      this.environmentList = list;
-      this.checkedList = [];
-      for (let i = 0; i < this.environmentList.length; i++) {
-        this.checkedList.push(false);
-      }
-    });
-  }
-
-  get environmentName() {
-    return this.environmentForm.get('environmentName');
-  }
-
-  get dockerImage() {
-    return this.environmentForm.get('dockerImage');
-  }
-
-  get name() {
-    return this.environmentForm.get('name');
-  }
-
-  get channels() {
-    return this.environmentForm.get('channels') as FormArray;
-  }
-
-  get dependencies() {
-    return this.environmentForm.get('dependencies') as FormArray;
-  }
-
-  initEnvForm() {
-    this.isVisible = true;
-    this.environmentName.reset();
-    this.dockerImage.reset();
-    this.name.reset();
-    this.channels.clear();
-    this.dependencies.clear();
-  }
-
-  checkStatus() {
-    return (
-      this.environmentName.invalid ||
-      this.dockerImage.invalid ||
-      this.name.invalid ||
-      this.channels.invalid ||
-      this.dependencies.invalid
-    );
-  }
-
-  closeModal() {
-    this.isVisible = false;
-  }
-
-  addChannel() {
-    this.channels.push(this.fb.control(null, Validators.required));
-  }
-
-  addDependencies() {
-    this.dependencies.push(this.fb.control(null, Validators.required));
-  }
-
-  deleteItem(arr: FormArray, index: number) {
-    arr.removeAt(index);
-  }
-
-  createEnvironment() {
-    this.isVisible = false;
-    const newEnvironmentSpec = this.createEnvironmentSpec();
-    console.log(newEnvironmentSpec);
-    this.environmentService.createEnvironment(newEnvironmentSpec).subscribe(
-      () => {
-        this.nzMessageService.success('Create Environment Success!');
-        this.fetchEnvironmentList();
-      },
-      (err) => {
-        this.nzMessageService.error(`${err}, please try again`, {
-          nzPauseOnHover: true,
-        });
-      }
-    );
-  }
-
-  createEnvironmentSpec() {
-    const environmentSpec = {
-      name: this.environmentForm.get('environmentName').value,
-      dockerImage: this.environmentForm.get('dockerImage').value,
-      kernelSpec: {
-        name: this.environmentForm.get('name').value,
-        channels: [],
-        dependencies: [],
-      },
-    };
-
-    for (const channel of this.channels.controls) {
-      environmentSpec.kernelSpec.channels.push(channel.value);
-    }
-
-    for (const dependency of this.dependencies.controls) {
-      environmentSpec.kernelSpec.dependencies.push(dependency.value);
-    }
-
-    return environmentSpec;
-  }
-
-  // TODO(kobe860219): Update an environment
-  updateEnvironment(id: string, data) {}
-
-  onDeleteEnvironment(name: string, onMessage: boolean) {
-    this.environmentService.deleteEnvironment(name).subscribe(
-      () => {
-        if (onMessage === true) {
-          this.nzMessageService.success('Delete Experiment Successfully!');
-        }
-        this.fetchEnvironmentList();
-      },
-      (err) => {
-        if (onMessage === true) {
-          this.nzMessageService.error(err.message);
-        }
-      }
-    );
-  }
-
-  deleteEnvironments() {
-    for (let i = this.checkedList.length - 1; i >= 0; i--) {
-      console.log(this.environmentList[i].environmentSpec.name);
-      if (this.checkedList[i] === true && this.environmentList[i].environmentSpec.name != 'notebook-env') {
-        this.onDeleteEnvironment(this.environmentList[i].environmentSpec.name, false);
-      }
-    }
-
-    this.selectAllChecked = false;
-  }
-
-  selectAllEnv() {
-    for (let i = 0; i < this.checkedList.length; i++) {
-      this.checkedList[i] = this.selectAllChecked;
-    }
-  }
+  ngOnInit() {}
 }
