+++ submarine-workbench/workbench-web/src/app/pages/workbench/environment/environment.module.ts
@@ -1,3 +1,22 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
@@ -9,16 +28,9 @@ import { EnvironmentService } from '@submarine/services/environment-services/env
 import { EnvironmentComponent } from './environment.component';
 import { EnvironmentListComponent } from './environment-home/environment-list/environment-list.component';
 import { EnvironmentFormComponent } from './environment-home/environment-form/environment-form.component';
-import { EnvironmentInfoComponent } from './environment-info/environment-info.component';
 
 @NgModule({
-  declarations: [
-    EnvironmentComponent,
-    EnvironmentHomeComponent,
-    EnvironmentListComponent,
-    EnvironmentFormComponent,
-    EnvironmentInfoComponent,
-  ],
+  declarations: [EnvironmentComponent, EnvironmentHomeComponent, EnvironmentListComponent, EnvironmentFormComponent],
   imports: [CommonModule, ReactiveFormsModule, FormsModule, RouterModule, NgZorroAntdModule, EnvironmentRoutingModule],
   providers: [EnvironmentService],
   exports: [EnvironmentComponent],
